Refactor: split code in files and add license notice
