+++ packages/cli/src/lib/transpile/index.ts
@@ -1,0 +1,2 @@
+export {transpileFile} from './file.js';
+export {transpileProject} from './project.js';
